+++ src/main/scala/com/programmaticallyspeaking/ncd/nashorn/PrintSupport.scala
@@ -1,7 +1,10 @@
 package com.programmaticallyspeaking.ncd.nashorn
 
 import com.programmaticallyspeaking.ncd.host.PrintMessage
+import com.programmaticallyspeaking.ncd.messaging.Observer
+import com.programmaticallyspeaking.ncd.nashorn.NashornDebuggerHost.{InternalState, Pause, Unpause}
 import com.sun.jdi.event.{BreakpointEvent, Event}
+import com.sun.jdi.request.BreakpointRequest
 import com.sun.jdi.{ArrayReference, ClassType, ThreadReference, Value}
 import org.slf4s.Logging
 
@@ -26,6 +29,12 @@ trait PrintSupport { self: NashornDebuggerHost with Logging =>
   def enablePrintCapture(global: ClassType): Unit = {
     Breakpoints.enableBreakingAtGlobalPrint(global) { breakpointRequests =>
       breakpointRequests.foreach(_.onEventDo(eventHandler))
+
+      internalStateSubject.subscribe(Observer.from[InternalState] {
+        case Pause   => breakpointRequests.foreach(_.disable())
+        case Unpause => breakpointRequests.foreach(_.enable())
+      })
+
     }
   }